chore: 식별자 타입 Integer -> Long으로 수정

+++ BE-MyCarMaster/src/main/java/softeer/bemycarmaster/api/domain/wheeldrive/controller/WheelDriveController.java
@@ -29,7 +29,7 @@ public class WheelDriveController {
 		@RequestBody @Valid GetWheelDrivesRequest getWheelDrivesRequest
 	) {
 
-		Integer trimId = getWheelDrivesRequest.getTrimId();
+		Long trimId = getWheelDrivesRequest.getTrimId();
 		GetWheelDrivesResponse getWheelDrivesResponse = getWheelDrivesUseCase.execute(trimId);
 		return Response.createSuccessResponse(getWheelDrivesResponse);
 	}
